Update main.tex
Updated the S_t formula for Bachelier to be consistent with the r != 0 formula. Added replication/proof up to deriving the chooser option formula for r != 0, haven't substituted the S_t = ... in yet. Did not write out the proof for special case r = 0 yet, though it the chooser option formula at the bottom is for r = 0.